css: productForm, productList, app

+++ src/components/ProductList.jsx
@@ -1,16 +1,36 @@
-function ProductList({ products, onDelete, onEdit }) {
+import "./ProductList.css"
+
+function ProductList({ products, onDelete, onEdit, loading }) {
+
+
+  if (loading) {
+    return <p className="spinner">Loading...</p>;
+  }
+
+  if ( !loading && products.length === 0) {
+    return <p style={{ marginTop: "20px", textAlign: "center" }}>No products found 📭.</p>;
+  }
+
+
   return (
-    <div>
+    <div className="product-flex">
       {products.map((product) => (
-        <div key={product.id}>
-          <p>{product.name}</p>
-          <p>Quantity: {product.quantity}</p>
-          <p>Price: {product.price}</p>
-          <p>Category: {product.category}</p>
 
-          <button onClick={() => onDelete(product.id)}>Delete</button>
-          <button onClick={() => onEdit(product)}>Edit</button>
+        <div className="product-card" key={product.id}>
 
+          <div className="product-info">
+            <strong>{product.name}</strong>
+            <span>Quantity: {product.quantity}</span>
+            <span>Price: {product.price}</span>
+            <span>Category: {product.category}</span>
+          </div>
+
+
+          <div className="product-actions">
+            <button onClick={() => onDelete(product.id)}>Delete</button>
+            <button onClick={() => onEdit(product)}>Edit</button>
+
+          </div>
 
           <hr />
         </div>
